rfctr: move htmlify_matrix_of_cell_texts()
We'll shortly be adding a group of objects which facilitate working with
the `<table>` HTML that appears in `.metadata.text_as_html`. These
objects will be used both by partitioning and chunking.

Create a new home for these and move `htmlify_matrix_of_cell_texts()`
here along with its tests, soon to be joined by the others.
